+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/AICommitAction.kt
@@ -3,6 +3,7 @@ package com.github.blarc.ai.commits.intellij.plugin
 import com.github.blarc.ai.commits.intellij.plugin.AICommitsBundle.message
 import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
 import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
+import com.github.blarc.ai.commits.intellij.plugin.settings.AppSettings
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.diff.impl.patch.IdeaTextPatchBuilder
@@ -14,6 +15,8 @@ import com.intellij.openapi.vcs.VcsDataKeys
 import com.intellij.openapi.vcs.changes.Change
 import com.intellij.openapi.vcs.ui.CommitMessage
 import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
+import com.knuddels.jtokkit.Encodings
+import com.knuddels.jtokkit.api.EncodingType
 import git4idea.repo.GitRepositoryManager
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
@@ -60,6 +63,7 @@ class AICommitAction : AnAction(), DumbAware {
                 try {
                     val generatedCommitMessage = openAIService.generateCommitMessage(diff, hint, 1)
                     commitMessageField.setCommitMessage(generatedCommitMessage)
+                    AppSettings.instance.recordHit()
                 } catch (e: Exception) {
                     commitMessageField.setCommitMessage("Error generating commit message")
                     sendNotification(Notification.unsuccessfulRequest(e.message ?: "Unknown error"))
@@ -104,5 +108,11 @@ class AICommitAction : AnAction(), DumbAware {
                 }
             }
             .joinToString("\n")
+    }
+
+    private fun isPromptTooLarge(prompt: String): Boolean {
+        val registry = Encodings.newDefaultEncodingRegistry()
+        val encoding = registry.getEncoding(EncodingType.CL100K_BASE)
+        return encoding.countTokens(prompt) > 4000
     }
 }